Split up implementation and added architecture.md
